requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
django==5.2.3
django-anymail==13.0
django-cors-headers==4.7.0
django-unfold==0.60.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
markdown==3.8.2
packaging==25.0
pillow==11.3.0
pyjwt==2.9.0
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
